Rename fx constuctors to avoid collisions
