+++ arrow-core-data/src/main/kotlin/arrow/core/Const.kt
@@ -5,9 +5,6 @@ import arrow.higherkind
 import arrow.typeclasses.Applicative
 import arrow.typeclasses.Semigroup
 import arrow.typeclasses.Show
-import arrow.typeclasses.suspended.BindSyntax
-import kotlin.coroutines.Continuation
-import kotlin.coroutines.intrinsics.suspendCoroutineUninterceptedOrReturn
 
 fun <A, T> ConstOf<A, T>.value(): A = this.fix().value()
 
@@ -52,38 +49,3 @@ fun <T, A, G> ConstOf<A, Kind<G, T>>.sequence(GA: Applicative<G>): Kind<G, Const
 
 inline fun <A> A.const(): Const<A, Nothing> =
   Const(this)
-
-object const {
-  fun <A, T> eager(c: suspend EagerBind<ConstPartialOf<A>>.() -> A): Const<A, T> {
-    val continuation: ConstContinuation<A, A> = ConstContinuation()
-    return continuation.startCoroutineUninterceptedAndReturn {
-      Const.just(c())
-    } as Const<A, T>
-  }
-
-  suspend operator fun <A, T> invoke(c: suspend BindSyntax<ConstPartialOf<A>>.() -> A): Const<A, T> =
-    suspendCoroutineUninterceptedOrReturn { cont ->
-      val continuation = ConstSContinuation(cont as Continuation<ConstOf<A, T>>)
-      continuation.startCoroutineUninterceptedOrReturn {
-        Const.just(c())
-      }
-    }
-
-  internal class ConstSContinuation<A, T>(
-    parent: Continuation<ConstOf<A, T>>
-  ) : SuspendMonadContinuation<ConstPartialOf<A>, T>(parent) {
-    override fun ShortCircuit.recover(): Const<A, T> =
-      throw this
-
-    override suspend fun <B> Kind<ConstPartialOf<A>, B>.bind(): B =
-      value() as B
-  }
-
-  internal class ConstContinuation<A, T> : MonadContinuation<ConstPartialOf<A>, T>() {
-    override fun ShortCircuit.recover(): Const<A, T> =
-      throw this
-
-    override suspend fun <B> Kind<ConstPartialOf<A>, B>.bind(): B =
-      value() as B
-  }
-}
