+++ arrow-core-data/src/main/kotlin/arrow/core/Eval.kt
@@ -1,10 +1,6 @@
 package arrow.core
 
-import arrow.Kind
 import arrow.higherkind
-import arrow.typeclasses.suspended.BindSyntax
-import kotlin.coroutines.Continuation
-import kotlin.coroutines.intrinsics.suspendCoroutineUninterceptedOrReturn
 
 fun <A> EvalOf<A>.value(): A = this.fix().value()
 
@@ -366,39 +362,4 @@ fun <A, B> Iterator<A>.iterateRight(lb: Eval<B>, f: (A, Eval<B>) -> Eval<B>): Ev
   fun loop(): Eval<B> =
     Eval.defer { if (this.hasNext()) f(this.next(), loop()) else lb }
   return loop()
-}
-
-object eval {
-  fun <A> eager(c: suspend EagerBind<ForEval>.() -> A): Eval<A> {
-    val continuation: EvalContinuation<A> = EvalContinuation()
-    return continuation.startCoroutineUninterceptedAndReturn {
-      Eval.just(c())
-    } as Eval<A>
-  }
-
-  suspend operator fun <A> invoke(c: suspend BindSyntax<ForEval>.() -> A): Eval<A> =
-    suspendCoroutineUninterceptedOrReturn { cont ->
-      val continuation = EvalSContinuation(cont as Continuation<EvalOf<A>>)
-      continuation.startCoroutineUninterceptedOrReturn {
-        Eval.just(c())
-      }
-    }
-
-  internal class EvalSContinuation<A>(
-    parent: Continuation<EvalOf<A>>
-  ) : SuspendMonadContinuation<ForEval, A>(parent) {
-    override fun ShortCircuit.recover(): Kind<ForEval, A> =
-      throw this
-
-    override suspend fun <A> Kind<ForEval, A>.bind(): A =
-      fix().value()
-  }
-
-  internal class EvalContinuation<A> : MonadContinuation<ForEval, A>() {
-    override fun ShortCircuit.recover(): Kind<ForEval, A> =
-      throw this
-
-    override suspend fun <A> Kind<ForEval, A>.bind(): A =
-      fix().value()
-  }
 }
