+++ arrow-core-data/src/main/kotlin/arrow/core/evalFx.kt
@@ -1,0 +1,41 @@
+package arrow.core
+
+import arrow.Kind
+import arrow.typeclasses.suspended.BindSyntax
+import kotlin.coroutines.Continuation
+import kotlin.coroutines.intrinsics.suspendCoroutineUninterceptedOrReturn
+
+object evalFx {
+  fun <A> eager(c: suspend EagerBind<ForEval>.() -> A): Eval<A> {
+    val continuation: EvalContinuation<A> = EvalContinuation()
+    return continuation.startCoroutineUninterceptedAndReturn {
+      Eval.just(c())
+    } as Eval<A>
+  }
+
+  suspend operator fun <A> invoke(c: suspend BindSyntax<ForEval>.() -> A): Eval<A> =
+    suspendCoroutineUninterceptedOrReturn { cont ->
+      val continuation = EvalSContinuation(cont as Continuation<EvalOf<A>>)
+      continuation.startCoroutineUninterceptedOrReturn {
+        Eval.just(c())
+      }
+    }
+
+  internal class EvalSContinuation<A>(
+    parent: Continuation<EvalOf<A>>
+  ) : SuspendMonadContinuation<ForEval, A>(parent) {
+    override fun ShortCircuit.recover(): Kind<ForEval, A> =
+      throw this
+
+    override suspend fun <A> Kind<ForEval, A>.bind(): A =
+      fix().value()
+  }
+
+  internal class EvalContinuation<A> : MonadContinuation<ForEval, A>() {
+    override fun ShortCircuit.recover(): Kind<ForEval, A> =
+      throw this
+
+    override suspend fun <A> Kind<ForEval, A>.bind(): A =
+      fix().value()
+  }
+}
